+++ src/compiler/program-ctx.ts
@@ -19,15 +19,16 @@ import {
 	ScopeVariableDeclaration,
 } from "./logic";
 import { KipperLogger } from "../logger";
-import { CompoundStatement, FunctionDefinition, RootFileParseToken, VariableDeclaration } from "./tokens";
+import { CompoundStatement, FunctionDeclaration, RootFileParseToken, VariableDeclaration } from "./tokens";
 import {
-	BuiltInOverwriteError,
-	DuplicateFunctionDefinitionError,
-	DuplicateVariableDefinitionError,
-	InvalidGlobalError,
-	UnknownFunctionDefinition,
-	UnknownTypeError,
-	UnknownVariableDefinition,
+  BuiltInOverwriteError,
+  FunctionDefinitionAlreadyExistsError,
+  IdentifierAlreadyUsedByFunctionError,
+  IdentifierAlreadyUsedByVariableError,
+  InvalidGlobalError,
+  UnknownFunctionIdentifier,
+  UnknownTypeError,
+  UnknownVariableIdentifier, VariableDefinitionAlreadyExistsError,
 } from "../errors";
 
 /**
@@ -58,7 +59,7 @@ export class CompileAssert {
 	 */
 	public functionIsDefined(identifier: string): void {
 		if (!this.programCtx.getGlobalFunction(identifier)) {
-			throw new UnknownFunctionDefinition(identifier);
+			throw new UnknownFunctionIdentifier(identifier);
 		}
 	}
 
@@ -68,7 +69,7 @@ export class CompileAssert {
 	 */
 	public variableIsDefined(identifier: string): void {
 		if (!this.programCtx.getGlobalFunction(identifier)) {
-			throw new UnknownVariableDefinition(identifier);
+			throw new UnknownVariableIdentifier(identifier);
 		}
 	}
 
@@ -78,30 +79,71 @@ export class CompileAssert {
 	 */
 	public functionIdentifierNotUsed(identifier: string): void {
 		if (this.programCtx.getGlobalFunction(identifier)) {
-			throw new DuplicateFunctionDefinitionError(identifier);
+			throw new IdentifierAlreadyUsedByFunctionError(identifier);
 		}
 	}
 
 	/**
 	 * Asserts that the passed variable identifier is not defined.
 	 * @param identifier The identifier of the variable.
-	 * @param scope The scope to use. If undefined, then it will use the global scope of the
-	 * {@link KipperProgramContext program ctx}.
+	 * @param scope The scope to also check besides the global scope. If undefined, then it will only the global scope
+   * of the {@link KipperProgramContext program}.
 	 */
-	public variableIdentifierNotUsed(identifier: string, scope?: CompoundStatement | KipperProgramContext): void {
-		let actualScope: ScopeDeclaration[];
-		if (scope === undefined) {
-			actualScope = this.programCtx.globalScope;
-		} else if (scope instanceof CompoundStatement) {
-			actualScope = scope.localScope;
-		} else {
-			actualScope = scope.globalScope;
+	public variableIdentifierNotUsed(identifier: string, scope?: CompoundStatement): void {
+    // Always check in the global scope
+    const check = (v: { identifier: string; }) => v instanceof ScopeVariableDeclaration && v.identifier === identifier;
+		if (this.programCtx.globalScope.find(check)) {
+			throw new IdentifierAlreadyUsedByVariableError(identifier);
 		}
 
-		if (actualScope.find((v) => v instanceof ScopeVariableDeclaration && v.identifier === identifier)) {
-			throw new DuplicateVariableDefinitionError(identifier);
-		}
-	}
+    if (this.programCtx.getGlobalVariable(identifier)) {
+      throw new IdentifierAlreadyUsedByVariableError(identifier);
+    }
+
+    // Also check in the local scope if it was passed
+    if (scope !== undefined && scope?.localScope.find(check)) {
+      throw new IdentifierAlreadyUsedByVariableError(identifier);
+    }
+  }
+
+  /**
+   * Asserts that the passed variable identifier has not been defined yet!
+   * @param identifier The identifier to check for in the global scope.
+   * @param scope The scope to also check besides the global scope. If undefined, then it will only the global scope
+   * of the {@link KipperProgramContext program}.
+   */
+  public variableIdentifierNotDefined(identifier: string, scope?: CompoundStatement): void {
+    // Always check in the global scope
+    const check = (v: { identifier: string; }) => {
+      // Return true only if the identifier match and the variable is DEFINED
+      return v instanceof ScopeVariableDeclaration && v.identifier === identifier && v.isDefined;
+    };
+
+    if (this.programCtx.globalScope.find(check)) {
+      throw new VariableDefinitionAlreadyExistsError(identifier);
+    }
+
+    // Also check in the local scope if it was passed
+    if (scope !== undefined && scope?.localScope.find(check)) {
+      throw new VariableDefinitionAlreadyExistsError(identifier);
+    }
+  }
+
+  /**
+   * Asserts that the passed variable identifier has not been defined yet!
+   * @param identifier The identifier to check for in the global scope.
+   */
+  public functionIdentifierNotDefined(identifier: string): void {
+    // Always check in the global scope
+    const check = (v: { identifier: string; }) => {
+      // Return true only if the identifier match and the function is DEFINED
+      return v instanceof ScopeFunctionDeclaration && v.identifier === identifier && v.isDefined;
+    };
+
+    if (this.programCtx.globalScope.find(check)) {
+      throw new FunctionDefinitionAlreadyExistsError(identifier);
+    }
+  }
 
 	/**
 	 * Asserts that the passed identifier does not exist as a built-in global.
@@ -435,10 +477,18 @@ export class KipperProgramContext {
 	/**
 	 * Adds a new declaration entry to the global scope.
 	 */
-	public addNewGlobalScopeEntry(token: VariableDeclaration | FunctionDefinition): void {
-		this.assert.builtInNotDefined(token.identifier);
-		this.assert.functionIdentifierNotUsed(token.identifier);
-		this.assert.variableIdentifierNotUsed(token.identifier);
+	public addNewGlobalScopeEntry(token: VariableDeclaration | FunctionDeclaration): void {
+    this.assert.builtInNotDefined(token.identifier);
+
+    // Check that the identifier is not used by some other definition and that there has not been a previous definition.
+    if (token instanceof VariableDeclaration) {
+      this.assert.functionIdentifierNotUsed(token.identifier);
+      this.assert.variableIdentifierNotDefined(token.identifier);
+    } else {
+      this.assert.variableIdentifierNotUsed(token.identifier);
+      this.assert.functionIdentifierNotDefined(token.identifier);
+    }
+
 		this._globalScope = this._globalScope.concat(
 			token instanceof VariableDeclaration ? new ScopeVariableDeclaration(token) : new ScopeFunctionDeclaration(token),
 		);