+++ src/index.ts
@@ -6,3 +6,16 @@
 export * from "./compiler/";
 export * from "./logger";
 export * from "./errors";
+
+// Global variables to identify the project
+
+// eslint-disable-next-line no-unused-vars
+export const name = "kipper";
+// eslint-disable-next-line no-unused-vars
+export const version = "0.2.2";
+// eslint-disable-next-line no-unused-vars
+export const author = "Luna Klatzer";
+// eslint-disable-next-line no-unused-vars
+export const license = "GPL-3.0";
+// eslint-disable-next-line no-unused-vars
+export const github = "https://github.com/Luna-Klatzer/Kipper";